fix(order,company,settlement,ai): int 산술·volatile 참조·죽은 널 분기를 정리한다
- ProductImage/WorkforceTrendPoint: int 로 계산한 뒤 long 으로 넓히던 자리를 long 산술로 (상한을 키우거나 극단값이 오면 조용히 넘친다, java:S2184)
- BusinessDayCalculator: static volatile 필드를 AtomicReference 로 — "참조 교체만 원자적"이라는 계약을 타입에 드러낸다 (java:S3077)
- AnthropicChatAdapter: 단발 호출 응답의 metadata/usage 널 분기는 Spring AI 계약상 실행될 수 없어 제거. 청크마다 usage 가 없을 수 있는 스트리밍 경로의 널 검사는 유지 (java:S2583)

+++ company-service/src/main/java/github/lms/lemuel/company/domain/WorkforceTrendPoint.java
@@ -30,7 +30,8 @@ public record WorkforceTrendPoint(YearMonth month, int headcount, BigDecimal est
 
     /** 연속 인접 월 지점 — 전월 스냅샷 대비 증감을 계산한다. */
     static WorkforceTrendPoint withChange(CompanyWorkforce snapshot, CompanyWorkforce previous) {
-        BigDecimal headcountChange = BigDecimal.valueOf(snapshot.headcount() - previous.headcount());
+        // long 산술 — int 뺄셈 후 넓히면 극단값에서 부호가 뒤집힌 증감이 나온다.
+        BigDecimal headcountChange = BigDecimal.valueOf((long) snapshot.headcount() - previous.headcount());
         BigDecimal previousSalary = previous.estimatedAnnualSalary().orElse(null);
         BigDecimal currentSalary = snapshot.estimatedAnnualSalary().orElse(null);
         BigDecimal salaryChange = previousSalary == null || currentSalary == null
